Refatorar lógica de atualização de usuário no Config.jsx para remover código desnecessário e atualizar o README
Atualizar caminhos das imagens no README.md

Corrigir erro ao deletar produto sem ID

Adicionar log para armazenar token no localStorage

Atualizar estilo de edição de linha no adminProducts.css

Adicionar busca de produto atualizado após criação

Deletar produto com sucesso

+++ Front-End/src/services/produtoAPI.jsx
@@ -102,6 +102,9 @@ export const updateProduto = async (id, produtoData) => {
 
 // Função para deletar um produto
 export const deleteProduto = async (id) => {
+    if (!id) {
+        throw new Error("ID do produto é necessário para deletar.");
+    }
     try {
         const response = await axios.delete(`${API_BASE_URL}/${id}`, {
             headers: {
